+++ docs/guides/release-notes.md
@@ -24,7 +24,7 @@ Track all ZPA Terraform provider's releases. New resources, features, and bug fi
 - Supported Terraform version: **v1.x**
 
 ### Bug Fixes
-- [PR #485](https://github.com/zscaler/terraform-provider-zpa/pull/485) - Fixed drift related to common flattening and expand functions for all v2 Access policy resources.
+- [PR #486](https://github.com/zscaler/terraform-provider-zpa/pull/486) - Fixed drift related to common flattening and expand functions for all v2 Access policy resources.
   ~> **NOTE** This fix does not affect existing configurations using the `v2` policy type.
   
 ## 3.33.1 (September, 8 2024)